feat: map in record detail; removed map in analytics
