+++ components/contents/record/RecordDetail.tsx
@@ -1,12 +1,15 @@
-import React, { useEffect } from "react";
-import { Text, View } from "react-native";
+import React, { useEffect, useState } from "react";
+import { Dimensions, Text, View } from "react-native";
 import { useRecordSingle } from "../../../data/records";
 import { LoaderScreen } from "react-native-ui-lib";
 import DonutChartContainer from "../analytics/DonutChartContainer";
+import MapView, { LatLng, Marker, PROVIDER_GOOGLE } from "react-native-maps";
 
 type RecordDetailProps = {
     id: number;
 }
+
+const { width, height } = Dimensions.get("window");
 
 export const RecordDetail: React.FC<RecordDetailProps> = ({ id }) => {
     const { isLoading, data, mutate } = useRecordSingle(id);
@@ -21,9 +24,42 @@ export const RecordDetail: React.FC<RecordDetailProps> = ({ id }) => {
         )
     }
 
+    const ASPECT_RATIO = width / (height / 3);
+    const LATITUDE_DELTA = 0.02;
+    const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
+    const INITIAL_LAT = Number(data.record.latitude);
+    const INITIAL_LNG = Number(data.record.longitude);
+    const INITIAL_POSITION = {
+        latitude: INITIAL_LAT,
+        longitude: INITIAL_LNG,
+        latitudeDelta: LATITUDE_DELTA,
+        longitudeDelta: LONGITUDE_DELTA,
+    };
+
     return (
-        <View style={{ flexDirection: 'column', flex: 1, justifyContent: 'center', alignItems: 'center', gap: 16 }}>
+        <View style={{ flexDirection: 'column', flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+            <View
+                style={{
+                    height: 200,
+                    width,
+                }}
+            >
+                <MapView
+                    provider={PROVIDER_GOOGLE}
+                    initialRegion={INITIAL_POSITION}
+                    style={{
+                        height: 200,
+                        borderRadius: 20
+                    }}
+                    pitchEnabled
+                    zoomEnabled
+                    scrollEnabled={false}
+                    rotateEnabled={false}
+                >
+                    <Marker coordinate={INITIAL_POSITION} />
+                </MapView>
+            </View>
             <DonutChartContainer analyticsData={[data.record.analytics]} />
-        </View>
+        </View >
     )
 };
